refactor: rename scripts from snake_case to kebab-case
Standardize script filenames to kebab-case convention:
- self_runner → self-runner
- loop_in_claude → loop-in-claude
- claude_stream_formatter → claude-stream-formatter

Update all references in deno.json, docs, tests, memory, and
reset-to-main.sh. Remove stale dev/prod tasks from deno.json.
Delete superseded documents/whiteboard.md (migration plan completed).
Add engine/AGENTS.md module documentation.
