+++ scripts/claude-stream-formatter.ts
