+++ scripts/self-runner.ts
